Migrate to OAuth
This commit changes the security settings so that the user should be
logged in to submit a talk. We use Oauth2 and Single Sign On with Github.

Our MVC test now requires authentication and we don't really want to
add OAuth there so we remove the import of our config and use a mock
user to login automatically.

Our submit template is still a bit basic since we need to tell again who
we are. We're going to fix that soon.