Reuse security information
This commit updates our model to have a User entity so that it is easier
to retrieve the submissions of a given speaker. Rather than asking the
user to provide those information we extract them from the context.

First, we need to automatically create a user in our system when a user
logs in for the first time. To do so, we create a custom
AuthoritiesExtractor that intializes a `User` entry if necessary.

This User Principal can be injected in any Controller with the
`@AuthenticationPrincipal` annotation.