Migrate data-oriented service layer to `DataJpaTest`
Our `SubmissionService` needs the data layer. We could mock it and
create a test that doesn't use Spring at all or we could reuse the data
oriented slice annotation.

This commit does the latter and imports the `SubmissionService` that
should be tested. This removes the need from an explicit configuration.
We could have implemented the same thing with a component scan filter:

```
@DataJpaTest(includeFilters = @ComponentScan.Filter(
       type=FilterType.ASSIGNABLE_TYPE, classes=SubmissionService.class))
```